Improve handling of illegal tokens, clean up.
The previous logic did not handle the situation where token.ILLEGAL
was returned together with an empty literal string, and panicked
because it performed an out of bounds slice operation (s[low:high],
where low > high).

Improve the code so that doesn't happen, and add test cases for it.

The token.ILLEGAL with empty literal string possibility was a bug in
the go/scanner package (see golang/go#28112). It has been fixed in
https://golang.org/cl/141337, and shouldn't happen again.

Vendor a copy of go/scanner with https://golang.org/cl/141337 applied.
It can and should be removed after Go 1.12 is released with that fix.

Clean up the code in general.