Use a copy of go/token from Go 1.11.1.
This change fixes the build for Go 1.10.4 and Go 1.9.7,
as well as when this package is used in module mode (where
the vendor directory is ignored).

The previous commit vendored a copy of go/scanner from Go 1.11.1
with a bugfix backported. The go/scanner of Go 1.11.1 requires
go/token from Go 1.11.1 to build. So use a copy of that.

Change from vendoring the standard library packages to copying
them into an internal folder. This is to help this package build
in module mode, which ignores vendor directory completely.

I've considered alternatives which included having two different
versions of the highlight_go code for Go 1.11 and older, but
it was hard to work around the bug, produce good output, and have
tests passing in all the permutations (Go 1.10 and older, Go 1.11,
future Go 1.12). That's why I opted to make it work using the newer
go/{scanner,token} packages that have fewer known bugs.

Fixes #1.